@@ -2,7 +2,7 @@
 
 namespace Espeon {
     public static class Constants {
-        public static readonly Color EspeonColour = new Color(0xd1a9dd);
+        public static readonly Color EspeonColour = new(0xd1a9dd);
 
         public const string EspeonAscii = @"▓█████   ██████  ██▓███  ▓█████  ▒█████   ███▄    █ 
 ▓█   ▀ ▒██    ▒ ▓██░  ██▒▓█   ▀ ▒██▒  ██▒ ██ ▀█   █ 
