@@ -1,20 +1,20 @@
-﻿using Disqord;
+﻿using System;
+using System.Collections.Concurrent;
+using System.Threading.Tasks;
+using Disqord;
 using Disqord.Bot.Prefixes;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
 using NUnit.Framework;
-using System;
-using System.Collections.Concurrent;
-using System.Threading.Tasks;
 
 namespace Espeon.Test {
     public class PrefixServiceTests {
         private static readonly ILogger<PrefixService> Logger = new NullLogger<PrefixService>();
         
         private static readonly Snowflake GuildId = 0L;
-        private static readonly TestGuild Guild = new TestGuild("PrefixServiceTests", GuildId);
+        private static readonly TestGuild Guild = new("PrefixServiceTests", GuildId);
         
         private GuildPrefixes _guildPrefixes;
         private IServiceProvider _provider;
